refactor: extract getFolderViewState and reuse useScrollToTop/useNeedsToWait
Both FolderViewBody implementations (legacy + virtualized) duplicated
the isInError/hasDataToShow/isLoading/isEmpty derivation. Extract it
into a pure helper and reuse the existing useScrollToTop/useNeedsToWait
hooks so the two bodies only differ in render output, not state logic.
